+++ packages/datepicker/src/NativeTimePicker.tsx
@@ -22,6 +22,10 @@ export type NativeTimePickerProps = {
    * @default false
    */
   disableLabelAnimation?: boolean;
+  /** Tekst eller ikon som kommer før inputfelter
+   * @default <ClockIcon />
+   */
+  prepend?: React.ReactNode;
 } & React.InputHTMLAttributes<HTMLInputElement>;
 
 export const NativeTimePicker = React.forwardRef<
@@ -37,6 +41,7 @@ export const NativeTimePicker = React.forwardRef<
       feedback,
       variant,
       disableLabelAnimation,
+      prepend = <ClockIcon />,
       ...rest
     },
     ref: React.Ref<HTMLInputElement>,
@@ -46,7 +51,7 @@ export const NativeTimePicker = React.forwardRef<
       <BaseFormControl
         style={style}
         className={className}
-        prepend={<ClockIcon inline />}
+        prepend={prepend}
         label={label}
         feedback={feedback}
         variant={variant}
@@ -66,12 +71,15 @@ export const NativeTimePicker = React.forwardRef<
 
 type NativeTimePickerBaseProps = {
   [key: string]: any;
-};
+} & React.DetailedHTMLProps<
+  React.InputHTMLAttributes<HTMLInputElement>,
+  HTMLInputElement
+>;
 
 const NativeTimePickerBase = React.forwardRef<
   HTMLInputElement,
   NativeTimePickerBaseProps
->(({ onChange, id, ...rest }, ref) => {
+>(({ onChange, id, value, ...rest }, ref) => {
   const contextVariant = useVariant();
   const currentVariant = rest.variant || contextVariant;
   const {
@@ -82,6 +90,14 @@ const NativeTimePickerBase = React.forwardRef<
   useOnMount(() => {
     setFiller && !isTimepickerFilled && setFiller(true);
   });
+
+  React.useEffect(() => {
+    if (value) {
+      setFiller && !isTimepickerFilled && setFiller(true);
+    } else {
+      setFiller && isTimepickerFilled && setFiller(false);
+    }
+  }, [value, setFiller, isTimepickerFilled]);
 
   const handleChange = (event: any) => {
     if (isFilled(event.target)) {
@@ -100,6 +116,7 @@ const NativeTimePickerBase = React.forwardRef<
       type="time"
       className="eds-form-control eds-native-date-picker"
       onChange={handleChange}
+      value={value}
       {...rest}
     />
   );