+++ packages/expand/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.2.25](https://bitbucket.org/enturas/design-system/compare/@entur/expand@3.2.24...@entur/expand@3.2.25) (2021-01-29)
+
+**Note:** Version bump only for package @entur/expand
+
+## [3.2.24](https://bitbucket.org/enturas/design-system/compare/@entur/expand@3.2.23...@entur/expand@3.2.24) (2021-01-20)
+
+**Note:** Version bump only for package @entur/expand
+
 ## [3.2.23](https://bitbucket.org/enturas/design-system/compare/@entur/expand@3.2.22...@entur/expand@3.2.23) (2021-01-13)
 
 ### Bug Fixes